Customer improvements
Adds initial wrapper class for Customer, as Fortnox::API::Customer, that links
the three subclasses together. The wrapper maps to the entity class by default,
but forwards some methods to the validator and repository as well.

Changes name for the customer entity class to Entity instead of Customer.

Removes the #to_json method on entity since it's the repositorys job to do the
proper JSON convertion. The same functionality can still be achived with
.to_hash.to_json anyway.

Some other small fixes.