Adds request dump to error message.
Adds the responses request object to the error message. In order to see if an
error is local or remote you have to evaluate the error against the actual
request. Case in point is if the get_access_token is called more than once,
causing the account to be locked in Fortnox's database, the error errouniously
says that access token or client secret is missing even though they are both
supplied in the request. A debugging developer would conclude that this was a
gem error even though the error is really at Fortnox and require a support call
to fix.