Fixes ustomer spec
Fixes the fact that the missing name attribute in the subject of the inclusion
in test case shadowed the fact that the test didn't pass. The inclusion test
failed due to the missing name attribute and the attribute name was present in
the inspect string since all the attribute names of the subject are present in
that string.