Restructures files
Models are keept under models, but loose the entities folder.

Validators and repositories are moved out into their own folders.

The main structure is that folders are pluralised, models and validators for
example, but module names are singular, Attribute and Repository for example.
The reasoning is that the folders hold multiple instances of each class, but the
module namespace, when use, is refering to a single instance of a class.