Validator shared context refactored
Shared example in shared validator context was not correct. It worked
on models that did not have required attributes, but with those having
required attributes, those were not set. Therefore, the shared example
now assumes that the passed model is valid. It then updates the model
and then checks that the validation is working as expected.

Since models are imutable, a new variable (updated_model) is used to
store the update.