Refactors the API base class
Improves exception raised on remote server error. Introduces a new exception
class to hold remote server exceptions to make it clear for people using the
library that these errors are not in the gem, but at the server side.

Splits the base_spec into an api_spec and a base_spec. Many of the former class
methods in base has become instance methods instead and the roles of the API
module vs the API::Base class is clearer. This spec change underlines that
shift.

Turns most of the former class methods on API::Base into instance methods and
removes a fair chunk of them all together.

Simplifies the code of the http request methods in base using an around wrapper
method and some define_method meta programming.

Changes how headers are handled.

Extracts two new modules to hold the supporting methods for different aspects of
the base class.

Adds test cases for the new exception.