Refactors repository
Extracts the keyword arguments from the base repository into an options object.

Updates customer repository to use the new options object.

Adds a MODEL constant to the top level repository classes so that the
`instantiate` method can be pulled up into base.

Adds an `only` method to the repository loaders. This will be used as the
central logic when implementing the filter methods. Filter methods will be
callable directly on a repository or meta class, such as:
`Fortnox::API::Invoice.unpaid` to get only unpaid invoices. These aliases will
then call `only` passing their own method name as the argument.

Aliases the `only` method on the Customer meta class.

Renames the internal `base_uri` property for repositories since it's not really
a base uri, it's the part of the uri that you add for the resource that the
repository handles. The new name, `uri`, attempts to reflect this better.