Customer validator spec uses shared context
When introducing shared context to the validators/customer_spec,
I found that validators for currency and country_code was
not working. These are separate attributes and are already validated.
For instance, if assigning 'A' to currency, it is still nil and no
validations are added. We must either change the attribute classes
and add the error there, or skip error messages for these attributes.

Additionally, validators context is updated with a shared examples
for validating inclusion of strings and the former shared examples
'validate inclusion of' is now called 'validate inclusion of number'.