Changes validators from singletons to instances
Changes the validator base to be a class instead of a module. The module
approach together with allowing for multiple validator blocks caused state to
persist between test runs since the violations, for example, was shared on the
module instance instead of a class instance.

Refactors the validation classes to inherit from the base class instead of
extending the module.

Refactors the base spec to pass with the new structure.

Refactors the constant validator to use the same structure.

Refactors the validation matcher to account for the fact that the described
class itself doesn't have the required methods any more.