Customer validator spec successfull
Shared example for 'validates inclusion of string' now accepts invalid_string.
Some attributes (for instance Currency and CountryCode) only stores
the two/three first given characters and throws the rest away.
The 'nonsence-string' used in the example were then actually valid.
The new parameter can therefore be used for these kind of attributes
to give a string that is known to be invalid.