OrderBase refactored to module
Invoice and Order has a lot in common. OrderBase was implemented to
hold the shared code between those classes. But when you think about
it, there are no real superclass/subclass relation between those
classes. Therefore, OrderBase is made a module and both Invoice and Order
includes this module in this commit.

From @d-Pixie:
"I'd go with a straight Ruby module and mix it in to be honest. This isn't
really a superclass/subclass relationship, order/offer/invoice are not logically
more specialised versions of the same base type the way chess is a subtype of
board games, they just share traits.

It's generally better to favour composition over inheritance when you can since
it gives you a lot more flexibility with your object graph. View the shared
properties as just that, common traits, and put them in a module that you
include in the other models."