Repository examples refactored
Fortnox no longer allows HTTP requests (HTTPS must be used).
Because of that, I was forced to rebuild the VCR cassetts, and when
doing that, I encountered some problems.

Firsly, repository specs are order dependent. One must first persist
a resource to be able to find it. Therefore, all repository specs
should be run with order: :defined.

Secondly, requests are recorded by VCR, but when the cassetts needs to be
updated, like now, requests are made to the real API. We have not yet
implemented delete (any resources like Invoices cannot be deleted),
so when fetching all Customers for instance, the response is no longer 1
customer but 39. Each time the cassetts are updated, this number will need to
be updated.

Shared example group for 'save' is refactored. Instead of always
fetching resource with id 1, it is now fetching the resource just
created in that spec. Resource with id 1 may not always exist.