Repository::Order spec test response with OrderRow
Before, the response when saving an Order were never checked. This
commit introduces some testing to see that the response contains
the values we are expecting. In fact, when writing these tests I
discovered an error in the OrderRow model, which is why the spec
is failing for the moment...