Adds dev gem rubocop and fixes obsolete config
Rubocop was used before as well in CodeClimate. Now, the gem is added as
development gem, which means you can run it standalone, for instance
when using Rubocop's autocorrect function.

When added the gem, rubocop complained about obsolete configuration
file, so I was forced to update it to be able to run the gem.