Adds require_attribute matcher
Instead of using shared examples for checking if a model really requires
a given attribute, a new strategy is used here: custom matchers.
I got the inspiration from Thoughtbot's excellent shoulda-matchers.

Shared examples are good for certain things, for instance check that a
subclass behaves like the parent. When including a shared example with
it_behaves_like{ ... }, it adds "behaves like" to the description. This
works with the former example with subclasses, but in this case,
where we want to assert that a class requires a special attribute to be
set, we do not want the description to be something like "Customer
behaves like required attribute :name when not set...". Instead, we want
something like "Customer should require attribute :name". Custom
matchers lets us do just this, which is an indicator that
shared_examples is not the right way to go in this particular example.

Using custom matchers for this particular use case also makes the specs
easier to read in my oppinion and it's easier to write the custom matcher
compared to a shared_example (again for this use case, shared_examples
are great for some tests).