WIP: Changes Dry::Struct typing to :schema
The typing used in Fortnox::API::Types::Model (constructor_type)
defines how Dry::Struct typing will work. Before, it was set to
:symbolize, which implements :weak typing (perform coercions and
silently fallback otherwise). We do not want it to silently fallback.
Because of that, we switched to :schema to get strict typing. This means
we will get exceptions when a value is incorrect.