Adds coercion to all types
Since we switched the struct constructor from `:symbolized` to `:schema` with
the introduction of dry-struct, that broke `:symbolized`, we lost coercion of
the values that we had before. This commit adds that coercion back as
constructors on the types. We will most likely refactor the implementation of
this later to use dry-validation instead.