Adds support for multiple access tokens
Since requests against Fortnox are rate limited on a per API user basis the
recommended, by Fortnox, solution is to use several API users and balance
requests over them.

This commit adds support for the FORTNOX_API_ACCESS_TOKEN environment to contain
several access tokens belonging to the same client secret. These tokens are
read, on per line, from the environment variable and then cyckled through when
the gem makes requests. When the end of the list is reached it simply wraps,
creating an endless list of tokens that are reused over and over.