Adds basic implementations for simple types
Simple types implemented straight in mappers.rb for all four base types as well
as Array and Hash.
Comments out a fair few things that Hannes had in there for this hack to work.
Adds tests for the basic types in mappers/base_spec.rb. They are not fantastic
but prove that it works, as far as it is right now ...

This is all just naïve implementations that might need to move and get fleshed
out. But they do allow for rudimentary JSON conversion of all our custom base
types.