Fixes #64. Hash diff no longer includes unique id
Before, our hash diff method always included unique_id attribute, even
though it was not changed. The reason was that this unique_id was needed
in Repository saver. Now, this logic is moved to our models. So
instead of fetching the unique id from the hash of the model, we now
fetch it directly from the model.

To fix this, it required savers to return entity instances instead of
hashes. This fixes #64.