Adds unique id and stub constants to model
The model base class now carries the unique_id method implementation and the
model class simply designates the attribute name of the unique id attribute as a
constant.

The minimal viable hash for a model class is now available as a STUB constant on
the model class. This stub is also used as the default inout to the new method,
allowing us to get a minimal viable instance of any model by simply calling
`.new`.