First try to improve Token rotation
Token rotation should be done on class level, not instance level.
See #88 for more details.

Instead of setting Access-Token both when initialized and after each
request, it is not only set BEFORE each request. This is to handle both
token rotation for subsequent requests with the same repository object
as well as requests with several repository objects.

Dummy VCR cassettes added.