Changes parameter name in shared ex group
Parameter name was required_attribute, refering to required attributes
(those Fortnox requires you to set). Before, that was always the case.
When implementing Article, we found out that we needed to pass some
additional arguments, not required ones, to the shared example group to
get it to work. That's why this commit renames the parameter to prevent
the programmer to think about required attributes. Instead, this
parameter should be used to pass additional attributes that is needed to
create the current Fortnox resource.