Merges API::Base with API::Repository::Base
The only thing API::Base did was to provide code for
API::Repository::Base. That is why these classes are merged into
API::Repository::Base in this commit. Now it's more obvious what the
code that was in API::Base does and that is's only used by repositories.

Also, the dependency to HTTParty is narrowed since it's clear only
repositories use it.

The API::Repository::Base now takes a model as an argument to "new".
Before, that base class was in fact abstract, it was not possible to
instantiate the class without subclassing it and provice a MODEL
constant. With this change, we can now test the base class in isolation.