Removes implicit include of repository helpers
Before, specs marked with :integration had the repository helper
included automatically. Now, this include is explicit instead,
making it possible to mark a spec as an integration spec without
including this specific helper (which is precicely what I need to
do now that the repository base class needs to renew tokens).